Fix linux compiling issues and other bugs

+++ codemp/game/q_shared.c
@@ -1001,7 +1001,7 @@ Q_vsnprintf: always appends a trailing '\0', returns number of characters writte
 or returns -1 on failure or if the buffer would be overflowed.
 ============
 */
-int Q_vsnprintf( char *dest, int size, const char *fmt, va_list argptr ) {
+int Q_vsnprintf( char *dest, size_t size, const char *fmt, va_list argptr ) {
         int ret;
  
 #ifdef _WIN32
